<pdftex in pdf mode>: load epstopdf (\AtEndOfPackage), thus converting .eps to .pdf by default (since we also have a restricted \write18 enabled by default, precisely so we could enable this).
This can be wrong!

If the .pdf is the source, rather than the .eps, users should put
 \newcommand{\DoNotLoadEpstopdf}{}
before even the \documentclass line of the
document, in case the class loads graphics packages.

More information about the epstopdf run will be in the .log file;
see http://ctan.org/pkg/epstopdf-pkg for more info about the
epstopdf package itself.  In turn, it calls the epstopdf script
(http://ctan.org/pkg/epstopdf).

The TeX code to do this was written by Heiko
Oberdiek and posted to the latex-l list on 4 Mar
2009 11:42:57 (http://article.gmane.org/gmane.comp.tex.latex.latex3/1475).